185_KeepChosenScenarioHandoffHonest review fix
